test(protocol): scope snapshots to batches
Refs #1655
