+++ tests/Dekaf.Tests.Unit/Protocol/Snapshots/RecordBatchWireFormatSnapshotTests.Record_batches_match_golden_wire_format_windows.verified.txt
